requirements.txt
fpdf==1.7.2
speedtest-cli==2.1.3
ping3==4.0.4